REPL Discard
Removed old, clunk user repl. Fixed some bugs with argparse
interface. Reworked all the code to reference the matrix/grid
as a series of rows (No more damn column inconsitency confusion)
Changed the print values of the grid for clarity and neatness.
Removed a lot of the fluff code that came with the repl
(Though there is still some more that could use cutting)
Moved the colors and screen clearing functionality to new Util.py
file to reduce duplicate code.